Abandon hope all ye who enter

+++ Triangle.h
@@ -29,7 +29,7 @@ public:
 	Plane GetPlane() const;
 	// TODO: better name
 	// Assuming that the triangle is not split by the plane(NAN), at which point this result is worthless, which side is it on, +/0/-
-	int GetRelation(const Plane& plane) const;
+	float GetRelation(const Plane& plane) const;
 
 	// Figure out if these should be inlined or not
 	inline bool ContainsPoint(const glm::vec3& point) const noexcept;
